Modified structure of code.
